The joined parameters for rainbow come back as a comma separated list with spaces in them, but when the template is viewed coming back from AWS the list has no spaces.  This gets the 2 in sync. (+1 squashed commit) Squashed commits: [9c94311] The joined parameters for rainbow come back as a comma separated list with spaces in them, but when the template is viewed coming back from AWS the list has no spaces.  This gets the 2 in sync.
The joined parameters for rainbow come back as a comma separated list with spaces in them, but when the template is viewed coming back from AWS the list has no spaces.  This gets the 2 in sync.

The joined parameters for rainbow come back as a comma separated list with spaces in them, but when the template is viewed coming back from AWS the list has no spaces.  This gets the 2 in sync.